OpenCTI_Report/requirements.txt
pycti
markdown
beautifulsoup4
python-docx